src/weather/pyproject.toml
[tool.poetry]
name = "weather"
version = "0.1.0"
description = ""
authors = ["Bryni Sanjez <brynisanjez@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
django-environ = "^0.9.0"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
